Refactor dashboard module into package with modular architecture
