+++ src/sockets/dashboard/routes.py
@@ -1,0 +1,97 @@
+"""
+HTTP API routes for dashboard data.
+
+Contains API endpoints for dashboard data access and CSV download functionality.
+"""
+
+import csv
+import io
+import logging
+from flask import jsonify, Response
+
+from src.config import app
+from src.models.quiz_models import Teams, Answers
+
+logger = logging.getLogger(__name__)
+
+@app.route('/api/dashboard/data', methods=['GET'])
+def get_dashboard_data():
+    try:
+        # Get all answers ordered by timestamp
+        with app.app_context():
+            all_answers = Answers.query.order_by(Answers.timestamp.asc()).all()
+        
+        answers_data = []
+        for ans in all_answers:
+            # Get team name for each answer
+            team = Teams.query.get(ans.team_id)
+            team_name = team.team_name if team else "Unknown Team"
+            
+            answers_data.append({
+                'answer_id': ans.answer_id,
+                'team_id': ans.team_id,
+                'team_name': team_name,
+                'player_session_id': ans.player_session_id,
+                'question_round_id': ans.question_round_id,
+                'assigned_item': ans.assigned_item.value,
+                'response_value': ans.response_value,
+                'timestamp': ans.timestamp.isoformat()
+            })
+        
+        return jsonify({'answers': answers_data}), 200
+    except Exception as e:
+        logger.error(f"Error in get_dashboard_data: {str(e)}", exc_info=True)
+        return jsonify({'error': 'An error occurred while retrieving dashboard data'}), 500
+
+@app.route('/download', methods=['GET'])
+def download_csv():
+    try:
+        # Get all answers ordered by timestamp
+        with app.app_context():
+            all_answers = Answers.query.order_by(Answers.timestamp.asc()).all()
+        
+        # Create CSV content in memory
+        output = io.StringIO()
+        writer = csv.writer(output)
+        
+        # Write CSV header
+        writer.writerow(['Timestamp', 'Team Name', 'Team ID', 'Player ID', 'Round ID', 'Question Item (A/B/X/Y)', 'Answer (True/False)'])
+        
+        # Write data rows
+        for ans in all_answers:
+            # Get team name for each answer
+            team = Teams.query.get(ans.team_id)
+            team_name = team.team_name if team else "Unknown Team"
+            
+            writer.writerow([
+                ans.timestamp.strftime('%m/%d/%Y, %I:%M:%S %p'),  # Format timestamp like JavaScript toLocaleString()
+                team_name,
+                ans.team_id,
+                ans.player_session_id,
+                ans.question_round_id,
+                ans.assigned_item.value,
+                ans.response_value
+            ])
+        
+        # Get the CSV content
+        csv_content = output.getvalue()
+        output.close()
+        
+        # Create response with appropriate headers for CSV download
+        response = Response(
+            csv_content,
+            mimetype='text/csv',
+            headers={
+                'Content-Disposition': 'attachment; filename=chsh-game-data.csv'
+            }
+        )
+        
+        return response
+        
+    except Exception as e:
+        logger.error(f"Error in download_csv: {str(e)}", exc_info=True)
+        return Response(
+            "An error occurred while generating the CSV file",
+            status=500,
+            mimetype='text/plain'
+        )
